setup_05/apps/api/requirements.txt
fastapi
uvicorn
pytest
drain3
chromadb
langchain
openai
pydantic
python-dotenv
